+++ lib/transaction.dart
@@ -10,7 +10,7 @@
 part of redis;
 
 class _WarningConnection {
-  noSuchMethod(_) => throw RedisRuntimeError("Transaction in progress. "
+  noSuchMethod(_) => throw RedisRuntimeException("Transaction in progress. "
       "Please complete Transaction with .exec");
 }
 
@@ -28,7 +28,8 @@ class Transaction extends Command {
 
   Future send_object(object) {
     if (transaction_completed) {
-      return Future.error(RedisRuntimeError("Transaction already completed."));
+      return Future.error(
+          RedisRuntimeException("Transaction already completed."));
     }
 
     Completer c = Completer();
@@ -36,10 +37,10 @@ class Transaction extends Command {
     super.send_object(object).then((msg) {
       if (msg.toString().toLowerCase() != "queued") {
         c.completeError(
-            RedisError("Could not enqueue command: " + msg.toString()));
+            RedisException("Could not enqueue command: " + msg.toString()));
       }
-    }).catchError((error) {
-      c.completeError(error);
+    }).catchError((Object error, StackTrace st) {
+      c.completeError(error, st);
     });
     return c.future;
   }
@@ -59,16 +60,14 @@ class Transaction extends Command {
         while (_queue.isNotEmpty) {
           _queue.removeFirst();
         }
-        // return Future.error(TransactionError("transaction error "));
-        throw TransactionError("transaction error ");
-        //return null;
+        throw TransactionException("transaction error");
       } else {
         if (list.length != _queue.length) {
           int? diff = list.length - _queue.length;
           //return
-          throw RedisRuntimeError(
-              "There was $diff command(s) executed during transcation,"
-              "not going trough Transation handler");
+          throw RedisRuntimeException(
+              "There was $diff command(s) executed during transaction, "
+              "not going through Transaction handler");
         }
         int len = list.length;
         for (int i = 0; i < len; ++i) {